Bump version to 1.2.0 for clarity
There is an old, outdated Text gem on rubygems.org. It's from the same
codebase, but hasn't been updated since 2007 and won't work on Ruby 1.9. For
reasons unknown, however, it has a higher version number. In order to avoid
confusion, make sure that text (lower case) is obviously the newest code.